Protocol changes: EntityId (into protocol_changes) (#774)
* New Subscribers EntityId behavior for clients and peers

* Improve routing logging

* New Queryables EntityId behavior for clients and peers

* Improve routing logging

* Use proper QueryableId in Session and AdminSpace

* Sessions use runtime Id generator to avoid collisions

* AdminSpace use runtime Id generator to avoid collisions

* Use proper ResponderId

* Define EntityId type

* Add source_eid to SourceInfo

* Update source_info_stack_size test

* Update source_info_stack_size test

* Introduce EntityGlobalId type

* Add id() function to Subscriber, Queryable and Publisher

* Add Publication::with_source_info() function

* Code format

* Remove ref to PR #703

* Fix doctests

* Add comments

* Remove comments

+++ zenoh/src/lib.rs
@@ -79,7 +79,7 @@ extern crate zenoh_core;
 #[macro_use]
 extern crate zenoh_result;
 
-pub(crate) type Id = usize;
+pub(crate) type Id = u32;
 
 use git_version::git_version;
 use handlers::DefaultHandler;
